TWriteBackCache: refactoring - move implementation details to a separate header
